+++ src/app/ServiceInventory/InstanceFormInput.test.tsx
@@ -9,6 +9,7 @@ describe('Instance Form Input', () => {
     { name: "bool_param", type: "bool", description: "a boolean attribute", modifier: "rw+", default_value_set: false },
     { name: "db_url", type: "string", description: "url for a database", modifier: "rw+", default_value_set: false },
     { name: "id", type: "number", description: "id of instance", modifier: "rw+", default_value_set: false },
+    { name: "opt_num", type: "number?", description: "optional number attribute", modifier: "rw+", default_value_set: false },
   ];
   const handleInputChange = (value, event) => { return; };
   it('Should render checkbox input for boolean parameters', () => {
@@ -23,13 +24,13 @@ describe('Instance Form Input', () => {
     const textInput = wrapper.find('input[type="text"]');
     expect(textInput.getDOMNode()).toBeTruthy();
   });
-  it('Should render url input for if string parameter is a url', () => {
+  it('Should render url input, if string parameter is a url', () => {
     const attributes = { name: "name", not_editable: "", bool_param: '', db_url: '' };
     const wrapper = mount(<InstanceFormInput attributeModels={attributeModels} attributeName="db_url" attributes={attributes} handleInputChange={handleInputChange} />);
     const urlInput = wrapper.find('input[type="url"]');
     expect(urlInput.getDOMNode()).toBeTruthy();
   });
-  it('Should show description if it\'s available', () => {
+  it('Should show description, if it\'s available', () => {
     const attributes = { name: "name", not_editable: "", bool_param: '', db_url: '' };
     const wrapper = mount(<InstanceFormInput attributeModels={attributeModels} attributeName="db_url" attributes={attributes} handleInputChange={handleInputChange} />);
     const description = wrapper.find('.pf-c-form__helper-text');
@@ -40,5 +41,11 @@ describe('Instance Form Input', () => {
     const wrapper = mount(<InstanceFormInput attributeModels={attributeModels} attributeName="id" attributes={attributes} handleInputChange={handleInputChange} />);
     const textInput = wrapper.find('input[type="number"]');
     expect(textInput.getDOMNode()).toBeTruthy();
+  });
+  it('Should render number input, even if parameter is optional', () => {
+    const attributes = { name: "name", not_editable: "", bool_param: '', id: '' };
+    const wrapper = mount(<InstanceFormInput attributeModels={attributeModels} attributeName="opt_num" attributes={attributes} handleInputChange={handleInputChange} />);
+    const numberInput = wrapper.find('input[type="number"]');
+    expect(numberInput.getDOMNode()).toBeTruthy();
   });
 });