+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Fix for diagnosing update_rejected state (#118)
 - Fix service instance error message style and position (#122)
 - Fix error when updating an instance in rejected state (#128)
+- Improve support of optional and number attributes (#130, #131)
 
 # Release 1.3.0 (2020-09-08)